requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.3.3
pytest-django==4.9.0